+++ .devcontainer/devcontainer.json
@@ -7,11 +7,13 @@
 	"workspaceFolder": "/tachi",
 	"postAttachCommand": "./dev/bootstrap.sh",
 	"updateRemoteUserUID": true,
-	"containerUser": "tachi",
+	// "containerUser": "tachi",
 	"customizations": {
 		"vscode": {
 			"settings": {},
-			"extensions": []
+			"extensions": [
+				"dbaeumer.vscode-eslint"
+			]
 		}
 	}
 }
