+++ client/Justfile
@@ -1,0 +1,12 @@
+[private]
+interactive:
+	-@cd ../ && just
+
+# Run the client interactively. Changes made to files will trigger reloads.
+start:
+	pnpm dev
+
+# Test that the client passes typechecking and linting.
+test:
+	pnpm typecheck
+	pnpm lint
