+++ dev/aliases.fish
@@ -2,5 +2,7 @@
 
 defnew g=git
 defnew gp=git push
+defnew gsw=git switch
+defnew gswc=git switch -c
 defnew gl=git pull
 defnew rr="exec fish"
