+++ dev/deb/README.md
@@ -1,0 +1,3 @@
+MongoDB tools aren't available in apt. I want to have mongodb-database-tools inside the container.
+
+I don't like having `curl | sh` inside containers, so I've saved the file to the repo.
