+++ dev/deb/install.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+
+for file in /tachi/dev/deb/*.deb; do
+	apt install -y "$file"
+done
