+++ seeds/.gitignore
@@ -1,6 +1,7 @@
 # Single time fired scripts
-scripts/single-time/*
-!scripts/single-time/.gitkeep
+scripts/personal/*
+!scripts/personal/README.md
+!scripts/personal/example.js
 
 # Rerunner output that I don't want to commit
 scripts/rerunners/bpi-output.json
