+++ seeds/scripts/single-use/README.md
@@ -1,0 +1,10 @@
+This is for your own personal scripts.
+
+Code and files placed in here will not get shared with others.
+
+**Use this for one-off things that NOBODY else will find interesting.**
+
+If there's a chance your script might be useful for someone else, or it might need to be
+re-ran in the future, **DO NOT PUT IT IN HERE!!**
+
+Everything else should go in `public`!
